Time: 0 ms (100.00%) | Memory: 19.7 MB (74.28%) - LeetSync

+++ 662-maximum-width-of-binary-tree/maximum-width-of-binary-tree.cpp
@@ -21,7 +21,7 @@ public:
             int s = q.size();
             long long l = 0;
             long long r = 0;
-            int mini=q.front().second;
+            long long mini=q.front().second;
             for (int i = 0; i < s; i++) {
                 auto [node, idx] = q.front();
                 q.pop();
